Converted extension to a greasemonkey script so that it can be used in firefox.
Added greasemonkey header to content.js.
Commented out line 99: node.onmousedown = null as it was causing the script to
stop executing. Not sure why this is happening yet - needs to be investigated.

To install this as a greasemonkey script in firefox simply install greasemonkey
then open frictionless.user.js in firefox and greasemonkey should prompt you
to add it as a user script.

Tested and confirmed it works with guardian and independent (uk) links. Needs
further testing to make sure other links work and the code that deals with
app requests also works.